Refactor: create useAuth hook and outsource ContextWrapper
